EZ Level Changer and start of player data holder
A prefab and script created that can be dropped into a scene to easily
change scene on contact.

Also a prefab and two scripts added for the actual storage of data for
player between scenes, will update position and rotation automatically
by default unless scene is the first one started on with a playerDataObj
inside it.

Two example scenes of using the scene changer, as demonstrated please
ensure scenes are in the Build Settings if trying to change scene!